refactor: Apply PR Feedback
