+++ src/main/kotlin/racing/dto/MoveResult.kt
@@ -1,5 +1,5 @@
-package step3.dto
+package racing.dto
 
-import step3.model.Car
+import racing.model.Car
 
 data class MoveResult(val car: Car, val position: Int)
