+++ src/main/kotlin/racing/dto/RacingRecord.kt
@@ -1,0 +1,7 @@
+package racing.dto
+
+data class RacingRecord(val trackResults: ArrayList<TrackResult> = ArrayList()) {
+    fun add(trackResult: TrackResult) {
+        trackResults.add(trackResult)
+    }
+}
