+++ src/main/kotlin/racing/util/NumberGenerator.kt
@@ -1,4 +1,4 @@
-package step3.util
+package racing.util
 
 fun interface NumberGenerator {
     fun generate(minValue: Int, maxValue: Int): Int
