+++ src/main/kotlin/racing/util/RandomGenerator.kt
@@ -1,4 +1,4 @@
-package step3.util
+package racing.util
 
 object RandomGenerator : NumberGenerator {
     override fun generate(minValue: Int, maxValue: Int): Int {
