+++ src/test/kotlin/racing/model/CarTest.kt
@@ -1,4 +1,4 @@
-package step3.model
+package racing.model
 
 import io.kotest.assertions.throwables.shouldThrow
 import io.kotest.core.spec.style.FunSpec
